requirements.txt
torch
torchvision
flask
nltk
Pillow
